Made some abstract methods virtual Renamed act-specific classes to better distinguish between them Some improvement to card battle damage buttons, update to API 2.14.5
